:recycle: Change to use concern module for common scope, method.
