+++ app/models/global_issue_template.rb
@@ -2,47 +2,40 @@ class GlobalIssueTemplate < ActiveRecord::Base
   include Redmine::SafeAttributes
   include Concerns::IssueTemplate::Common
   unloadable
-  belongs_to :author, class_name: 'User', foreign_key: 'author_id'
-  belongs_to :tracker
-  validates :title, presence: true
-  validates :tracker, presence: true
-  validates_uniqueness_of :title, scope: :tracker_id
-  acts_as_list scope: :tracker
 
+  validates_uniqueness_of :title, scope: :tracker_id
   has_and_belongs_to_many :projects
 
-  # author and project should be stable.
   safe_attributes 'title',
                   'description',
                   'tracker_id',
                   'note',
                   'enabled',
+                  'is_default',
                   'issue_title',
                   'project_ids',
                   'position',
                   'author_id'
-  attr_accessible :title, :tracker_id, :issue_title, :description, :note,
-                  :enabled, :project_ids, :position, :author_id
 
+  attr_accessible :title, :tracker_id, :issue_title, :description, :note,
+                  :is_default, :enabled, :author, :project, :position
+
+  # for intermediate table assosciations
   scope :search_by_project, lambda { |project_id|
     joins(:projects).where(projects: { id: project_id })
   }
 
-  def checklist
-    #
-    # TODO: Exception handling
-    #
-    return [] if checklist_json.blank?
-    JSON.parse(checklist_json)
+  module Config
+    JSON_OBJECT_NAME = 'global_issue_template'.freeze
   end
+  Config.freeze
 
-  def template_json
-    result = attributes
-    result[:checklist] = checklist
-    result.delete('checklist_json')
-    template = {}
-    template[:global_issue_template] = result
-    template.to_json(root: true)
+  #
+  # In case set is_default and updated, others are also updated.
+  #
+  def check_default
+    return unless is_default? && is_default_changed?
+    self.class.search_by_tracker(tracker_id).update_all(is_default: false)
   end
 
   #
