+++ app/models/issue_template.rb
@@ -3,14 +3,8 @@ class IssueTemplate < ActiveRecord::Base
   include Concerns::IssueTemplate::Common
   unloadable
   belongs_to :project
-  belongs_to :author, class_name: 'User', foreign_key: 'author_id'
-  belongs_to :tracker
-  before_save :check_default
   validates :project_id, presence: true
-  validates :title, presence: true
-  validates :tracker, presence: true
   validates_uniqueness_of :title, scope: :project_id
-  acts_as_list scope: :tracker
 
   # author and project should be stable.
   safe_attributes 'title', 'description', 'tracker_id', 'note', 'enabled', 'issue_title', 'is_default',
@@ -19,35 +13,21 @@ class IssueTemplate < ActiveRecord::Base
                   :is_default, :enabled, :enabled_sharing, :author, :project, :position
 
   scope :enabled_sharing, -> { where(enabled_sharing: true) }
-  scope :is_default, -> { where(is_default: true) }
-  scope :not_default, -> { where(is_default: false) }
   scope :search_by_project, lambda { |prolect_id|
     where(project_id: prolect_id)
   }
+
+  module Config
+    JSON_OBJECT_NAME = 'issue_template'.freeze
+  end
+  Config.freeze
 
   #
   # In case set is_default and updated, others are also updated.
   #
   def check_default
     return unless is_default? && is_default_changed?
-    IssueTemplate.search_by_project(project_id).search_by_tracker(tracker_id).update_all(is_default: false)
-  end
-
-  def checklist
-    #
-    # TODO: Exception handling
-    #
-    return [] if checklist_json.blank?
-    JSON.parse(checklist_json)
-  end
-
-  def template_json
-    result = attributes
-    result[:checklist] = checklist
-    result.delete('checklist_json')
-    template = {}
-    template[:issue_template] = result
-    template.to_json(root: true)
+    self.class.search_by_project(project_id).search_by_tracker(tracker_id).update_all(is_default: false)
   end
 
   #
